[MacOS] Implement ScrollViewer
ScrollContentPresenter now inherits from NSScrollView. Layout calculations are shared with iOS, though it's not obvious to what extent this makes sense since the child model is rather different.
