+++ src/Uno.UI/UI/Xaml/IHasSizeThatFits.macOS.cs
@@ -1,0 +1,9 @@
+using CoreGraphics;
+
+namespace Uno.UI
+{
+	public interface IHasSizeThatFits
+	{
+		CGSize SizeThatFits(CGSize size);
+	}
+}
